Upload .stl files
These files comprise the novel magnetic socket I designed for my
Cerberus. This is a zero-lash virtual socket, where they three neodymium
magnets are epoxied into a printed carrier that orients them in three
axes. Skyved Teflon tape is cut with a hole punch and applied as a
bearing surface to one end of each magnet, where it attaches to the 3/8
inch ball.

For the rod end assemblies, I tried many things (as did many others) to
get a ball bearing concentric to the threaded adapter. Every approach I
tried that involved heat or chemical etching destroyed the surface of
the bearings. I ultimately settled on M3 hex head bolts in stainless
steel, tops of the heads faced smooth, then center-drilled on the lathe
with a small chamfer. I epoxied the ball into the chamfer on the bolt
head with JB Weld, with a fillet from the ball to the corners of the
hex. I'm pretty happy with the results.

The result is a dry joint, with the chrome, case hardened ball riding on
Teflon. There is great repeatability in these joints, as the magnets are
held in position against the ball while they are epoxied into the
carrier. I made some extra rod ends and some printed stands so that I
could glue up sets of these at a time.